add TaskCardController

+++ src/main/java/com/thoughtworks/nho/olsapi/repository/TaskCardRepository.java
@@ -3,5 +3,5 @@ package com.thoughtworks.nho.olsapi.repository;
 import com.thoughtworks.nho.olsapi.model.TaskCard;
 import org.springframework.data.repository.CrudRepository;
 
-public interface TaskCardRepository extends CrudRepository<TaskCard, Long> {
+public interface TaskCardRepository extends CrudRepository<TaskCard, Integer> {
 }
